+++ CHANGELOG.md
@@ -32,6 +32,7 @@ so make sure you follow the template
 * Add documentation
 * Add support for django 3.2
 * Add explicit support for python 3.6-3.9
+* Add `CSV_PERMISSIONS_STRICT` setting
 
 ## 0.1.0 2020-11-25